+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [0.2.4]
 - inter canister DB query calls working
 - removed QueryRow because it's effectively the same thing as DataRow
+- changed enum_value to use ArgNumber not i32
 
 ## [0.2.3]
 - breaking change to database API.  Queries are now constructed separately, the fluid method